+++ app/src/main/res/values/attrs.xml
@@ -35,6 +35,7 @@
     <attr name="ic_import_export" format="reference"/>
     <attr name="ic_save" format="reference"/>
     <attr name="ic_backup" format="reference"/>
+    <attr name="ic_add" format="reference"/>
 
     <!-- Can't refer to colors directly into drawable's xml-->
     <attr name="toolbar_shadow_drawable" format="reference"/>
